fix: collect all job responses
Loops over the activation stream until the stream is closed. An consumer was added to get intermediate jobs. In the ActivateJobsCommand the consumer is used to collect all jobs in one response. This is not necessary in the worker, here we want to already work on jobs if they are available. This need to be done as follow up