refactor: handle jobs while stream is open
Refactor job worker such that it uses the JobActivator and the request directly. This makes it possible to register an own consumer which will be called if a new response was incoming. This improved the performance of the job handling. Make the consumer async, which allows to be used in the job worker with the data pipeline. Add a task in the ActivateJobsCommand, which just collects all responses.

Remove the retry from the activation, since the activation is retried in
the worker after the poll interval.