feat: migrate to GrpcChannel
In order to use the dotnet native grpc.net channel we have to migrate
to the new GrpcChannel. Previously, we still used the Grpc.Core
channel which is backed with C code (non-native dotnet).

There are several changes which have to be done in order to migrate to
the new GrpcChannel. Important part is that ChannelOptions and
Credentials doesn't work as before.

1. GrpcChannel uses internally HttpClient, which means we have to set
   the keepAlive on the SocketsHttpHandler.
2. User-agent header needs to be set via an interceptor.

Co-authored-by: legalles <legalles@kappaeng.com>