feat: migrate to new SSLOptions
With the Grpc.net migration and the GrpcChannel we no longer can use the SslCredentials
to set our certificate. This is kept for backwards compatibility, but will not
accept any custom certificate. If we want specifiy a certificate we have to create a
X509Certificate2 and set it on the SocketsHttpHandler in the specific SslOptions.

Furthermore to allow custom self-signed certs we need to adjust the validation, otherwise
the certificates are rejected.